+++ src/Microsoft.Identity.Web.Certificateless/CertificatelessConstants.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Microsoft.Identity.Web.Certificateless
+{
+    internal class CertificatelessConstants
+    {
+        // Managed Identity Federated Identity Credential
+        internal const string DefaultTokenExchangeUrl = "api://AzureADTokenExchange";
+    }
+}